requirements.txt
apturl==0.5.2
bcrypt==3.2.0
bidict==0.22.1
blessings==1.7
blinker==1.6.2
bpython==0.22.1
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
curtsies==0.3.10
cwcwidth==0.1.6
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
fasteners==0.14.1
Flask==2.3.3
Flask-SocketIO==5.3.6
future==0.18.2
greenlet==1.1.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pygame==2.5.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
typing-extensions==3.10.0.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
usb-creator==0.3.7
wadllib==1.3.6
watchdog==2.1.6
websocket-client==1.6.2
Werkzeug==2.3.7
xdg==5
xkit==0.0.0
zipp==1.0.0
